refactor!: separate filter into its own directory
