+++ raillabel/filter/_filter_classes/_filter_abc.py
@@ -4,8 +4,8 @@
 import typing as t
 from abc import ABC, abstractmethod, abstractproperty
 
-from ..format._object_annotation import _ObjectAnnotation
-from ..format.frame import Frame
+from ...format._object_annotation import _ObjectAnnotation
+from ...format.frame import Frame
 
 
 class _FilterABC(ABC):
