+++ tests/test_raillabel/filter/test_filter.py
@@ -7,7 +7,7 @@ from pathlib import Path
 
 import pytest
 
-sys.path.insert(1, str(Path(__file__).parent.parent.parent))
+sys.path.insert(1, str(Path(__file__).parent.parent.parent.parent))
 
 import raillabel
